requirements.txt
Flask==2.3.3
supabase>=2.27.0,<3
python-dotenv==1.0.0
Werkzeug>=3.0.0
gunicorn==21.2.0
websockets>=14.0
Flask-Babel>=3.0.0,<4
Babel>=2.14,<3
